+++ Software/src/devboard/utils/timer.cpp
@@ -1,11 +1,11 @@
 #include "timer.h"
 
-MyTimer::MyTimer(unsigned long interval) : interval(interval) {
+MyTimer::MyTimer(uint32_t interval) : interval(interval) {
   previous_millis = millis();
 }
 
 bool MyTimer::elapsed(void) {
-  unsigned long current_millis = millis();
+  uint32_t current_millis = millis();
   if (current_millis - previous_millis >= interval) {
     previous_millis = current_millis;
     return true;
@@ -17,7 +17,7 @@ void MyTimer::reset(void) {
   previous_millis = millis();
 }
 
-void MyTimer::set_interval(unsigned long interval) {
+void MyTimer::set_interval(uint32_t interval) {
   this->interval = interval;
   reset();
 }
